# openapi.yaml
openapi: 3.0.3
info:
  title: Manufacturing Execution System API
  description: |
    Comprehensive MES API for jet engine component manufacturing.
    
    This API provides endpoints for managing work orders, quality control, 
    material tracking, equipment management, and comprehensive traceability 
    for aerospace manufacturing operations.
    
    ## Authentication
    All endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    API requests are limited to 1000 requests per hour per user.
    
    ## Error Handling
    The API uses standard HTTP status codes. Error responses include:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found
    - `409` - Conflict (business rule violation)
    - `422` - Unprocessable Entity (business validation error)
    - `500` - Internal Server Error
  version: 1.0.0
  contact:
    name: MES API Support
    email: support@mes-system.com
  license:
    name: Proprietary
servers:
  - url: https://api.mes-system.com/v1
    description: Production server
  - url: https://staging-api.mes-system.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /workorders:
    get:
      tags:
        - Work Orders
      summary: List work orders
      description: Retrieve a paginated list of work orders with filtering options
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [CREATED, RELEASED, IN_PROGRESS, COMPLETED, CANCELLED]
        - name: partNumber
          in: query
          schema:
            type: string
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: dueDateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dueDateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Work orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkOrder'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Work Orders
      summary: Create work order
      description: Create a new production work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkOrderRequest'
      responses:
        '201':
          description: Work order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /workorders/{workOrderId}:
    get:
      tags:
        - Work Orders
      summary: Get work order
      description: Retrieve a specific work order by ID
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Work order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Work Orders
      summary: Update work order
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkOrderRequest'
      responses:
        '200':
          description: Work order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'

    delete:
      tags:
        - Work Orders
      summary: Delete work order
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Work order deleted successfully
  /workorders/{workOrderId}/release:
    post:
      tags:
        - Work Orders
      summary: Release work order to production
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Work order released successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable entity - business validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    WorkOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workOrderNumber:
          type: string
        partId:
          type: string
          format: uuid
        partNumber:
          type: string
        partName:
          type: string
        routeId:
          type: string
          format: uuid
        quantityOrdered:
          type: integer
          minimum: 1
        quantityCompleted:
          type: integer
          minimum: 0
        quantityScr [continued...]apped:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [CREATED, RELEASED, IN_PROGRESS, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        customerOrder:
          type: string
        dueDate:
          type: string
          format: date
        scheduledStartDate:
          type: string
          format: date
        actualStartDate:
          type: string
          format: date
        scheduledEndDate:
          type: string
          format: date
        actualEndDate:
          type: string
          format: date
        siteId:
          type: string
          format: uuid
        operations:
          type: array
          items:
            $ref: '#/components/schemas/WorkOrderOperation'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkOrderOperation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        operationNumber:
          type: integer
        operationCode:
          type: string
        operationName:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, SKIPPED]
        quantityCompleted:
          type: integer
        quantityScr [continued...]apped:
          type: integer
        setupStartTime:
          type: string
          format: date-time
        setupEndTime:
          type: string
          format: date-time
        runStartTime:
          type: string
          format: date-time
        runEndTime:
          type: string
          format: date-time
        operatorId:
          type: string
          format: uuid
        workCenterId:
          type: string
          format: uuid
        equipmentId:
          type: string
          format: uuid
    CreateWorkOrderRequest:
      type: object
      required:
        - partNumber
        - quantityOrdered
        - siteId
      properties:
        partNumber:
          type: string
        quantityOrdered:
          type: integer
          minimum: 1
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
          default: NORMAL
        customerOrder:
          type: string
        dueDate:
          type: string
          format: date
        siteId:
          type: string
          format: uuid
    UpdateWorkOrderRequest:
      type: object
      properties:
        quantityOrdered:
          type: integer
          minimum: 1
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        customerOrder:
          type: string
        dueDate:
          type: string
          format: date
        scheduledStartDate:
          type: string
          format: date
        scheduledEndDate:
          type: string
          format: date
...